refactor(demo): extract shared helpers into vultron/demo/helpers/
Implements #489 — extract scenario-agnostic helper functions from the
2050-line two_actor_demo.py into a structured helpers package so they
can be reused across future demo scenarios.

New package: vultron/demo/helpers/
  - polling.py   — _poll_until generic primitive + all wait_for_* helpers
  - actions.py   — actor_notifies_* and actor_closes_case wrappers
  - seeding.py   — _dl_key, get_actor_by_id, seed_containers
  - sync.py      — SYNC-2 helpers: trigger_log_commit, verify_replica_state
  - verification.py — assertion primitives used by milestone verifiers
  - __init__.py  — full public re-export surface from all sub-modules

vultron/demo/scenario/two_actor_demo.py:
  - Re-imports all moved functions via helpers package (noqa: F401)
  - Retains reset_containers (patch target for TestResetContainers)
  - Retains finder_submits_report, vendor_validates_report, Q&A helpers,
    find_case_for_offer, verify_vendor_case_state, verify_case_actor_unused,
    all verify_m*_state milestone verifiers, run_two_actor_demo, main
  - Reduced from ~2050 lines to ~1210 lines

All 2592 tests pass; mypy, pyright, black, flake8 clean.
