+++ vultron/demo/helpers/seeding.py
@@ -1,0 +1,129 @@
+#  Copyright (c) 2026 Carnegie Mellon University and Contributors.
+#  - see Contributors.md for a full list of Contributors
+#  - see ContributionInstructions.md for information on how you can Contribute to this project
+#  Vultron Multiparty Coordinated Vulnerability Disclosure Protocol Prototype is
+#  licensed under a MIT (SEI)-style license, please see LICENSE.md distributed
+#  with this Software or contact permission@sei.cmu.edu for full terms.
+#  Created, in part, with funding and support from the United States Government
+#  (see Acknowledgments file). This program may include and/or can make use of
+#  certain third party source code, object code, documentation and other files
+#  ("Third Party Software"). See LICENSE.md for more details.
+#  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
+#  U.S. Patent and Trademark Office by Carnegie Mellon University
+
+"""Container-seeding and actor-lookup helpers for demo workflows.
+
+Provides :func:`_dl_key` for safe DataLayer path encoding,
+:func:`get_actor_by_id` for actor look-ups, and :func:`seed_containers` for
+the two-phase seeding sequence used by multi-actor scenarios.
+"""
+
+import logging
+from typing import Tuple
+from urllib.parse import quote
+
+from vultron.demo.utils import DataLayerClient, seed_actor
+from vultron.wire.as2.vocab.base.objects.actors import as_Actor
+
+logger = logging.getLogger(__name__)
+
+
+def _dl_key(key: str) -> str:
+    """URL-encode a DataLayer key for safe embedding in an API path segment.
+
+    Encodes characters that are illegal in URL path segments (e.g., colons in
+    URN-style keys like ``urn:uuid:...``).
+
+    Note: HTTP URL-based participant IDs (which contain literal slashes)
+    cannot be fetched via the single-segment ``/{key}`` DataLayer route even
+    after URL-encoding, because Starlette decodes ``%2F`` back to ``/`` before
+    path matching.  Such IDs must be handled via exception catching at the
+    call site.
+    """
+    return quote(str(key), safe="")
+
+
+def get_actor_by_id(client: DataLayerClient, actor_id: str) -> as_Actor:
+    """Fetch an actor record from a container by its full URI.
+
+    Args:
+        client: DataLayerClient connected to the target container.
+        actor_id: Full URI of the actor to fetch.
+
+    Returns:
+        The ``as_Actor`` object.
+
+    Raises:
+        ValueError: If the actor is not found.
+    """
+    actors = client.get("/actors/")
+    for a in actors:
+        actor = as_Actor.model_validate(a)
+        if actor.id_ == actor_id:
+            return actor
+    raise ValueError(f"Actor {actor_id!r} not found in container")
+
+
+def seed_containers(
+    finder_client: DataLayerClient,
+    vendor_client: DataLayerClient,
+    reporter_actor_id: str | None = None,
+    vendor_actor_id: str | None = None,
+) -> Tuple[as_Actor, as_Actor]:
+    """Seed both containers: create actor records and register cross-container peers.
+
+    The seeding is done in two phases to avoid ordering issues:
+
+    1. Create the local actor on each container independently.
+    2. Register each actor as a known peer on the other container.
+
+    This function is idempotent: re-running it returns existing actors
+    unchanged (the ``POST /actors/`` endpoint is idempotent).
+
+    Args:
+        finder_client: DataLayerClient connected to the Finder container.
+        vendor_client: DataLayerClient connected to the Vendor container.
+        reporter_actor_id: Optional deterministic URI for the Finder actor.
+            When absent the server derives one from ``VULTRON_BASE_URL``.
+        vendor_actor_id: Optional deterministic URI for the Vendor actor.
+            When absent the server derives one from ``VULTRON_BASE_URL``.
+
+    Returns:
+        Tuple of ``(finder, vendor)`` ``as_Actor`` objects as created on their
+        respective containers.
+    """
+    logger.info("Phase 1: creating local actors on each container...")
+    finder = seed_actor(
+        client=finder_client,
+        name="Finder",
+        actor_type="Person",
+        actor_id=reporter_actor_id,
+    )
+    logger.info("Finder actor seeded on Finder container: %s", finder.id_)
+
+    vendor = seed_actor(
+        client=vendor_client,
+        name="Vendor",
+        actor_type="Organization",
+        actor_id=vendor_actor_id,
+    )
+    logger.info("Vendor actor seeded on Vendor container: %s", vendor.id_)
+
+    logger.info("Phase 2: registering cross-container peers...")
+    seed_actor(
+        client=finder_client,
+        name="Vendor",
+        actor_type="Organization",
+        actor_id=vendor.id_,
+    )
+    logger.info("Vendor peer registered on Finder container: %s", vendor.id_)
+
+    seed_actor(
+        client=vendor_client,
+        name="Finder",
+        actor_type="Person",
+        actor_id=finder.id_,
+    )
+    logger.info("Finder peer registered on Vendor container: %s", finder.id_)
+
+    return finder, vendor
